+++ .github/CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
  - Added an editor-only window to quickly retrieve an OAuth access token (menu : `Blizzard API/Get Access Token`)
    - This is mostly useful in Blizzard's documentation, by providing an access token to test endpoints
   
+## [0.41.0] - 2020-05-01
+### Added
+ - Added support for the following WoW Profile endpoint category :
+   - Character Mythic Keystone Profile
+  
 ## [0.40.0] - 2020-05-01
 ### Added
  - Added support for the following WoW Profile endpoint category :